+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - A missing `init-ns` (e.g. `user`) no longer prints a stacktrace.
 
+### Changed
+
+- Switched from `clojure.tools.nrepl.*` to `nrepl.*`.
+
 ## 0.2.0
 
 Initial release